refactor(ln_voice_over): fold CLEAN into PARSE; drop cleaned/ directory
Stage 2 was 126 lines of regex cleanup whose only consumer was parse.
Collapsing the two drops an on-disk directory, a CLI command, a README
section, and a node from the pipeline diagram — matches the
"fewer steps / clearer file tree" goal from the Phase 2 audit.
Downstream stages renumbered to keep the sequence gap-free.
